% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuik-geo-data.R
\name{geo_data}
\alias{geo_data}
\title{Get List of All Databases for a given Theme from TUIK}
\usage{
geo_data(level = NULL, variable = NULL)
}
\arguments{
\item{level}{NUTS Level (2, 3, or 4)}

\item{variable}{Data Series Name}
}
\value{
A data tibble
}
\description{
Get List of All Databases for a given Theme from TUIK
}
\examples{
\dontrun{
geo_data(level = 2)
}

}
