% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuik-geo-map.R
\name{geo_map}
\alias{geo_map}
\title{Get Maps from TUIK}
\usage{
geo_map(level = c(2, 3, 4, 9), dataframe = FALSE)
}
\arguments{
\item{level}{NUTS Level (2, 3, 4, or 9)}

\item{dataframe}{Return data as data frame}
}
\value{
An sf object or tibble
}
\description{
Get Maps from TUIK
}
\examples{
\dontrun{
geo_map(level = 2)
}

}
