% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuik-statistical-databases.R
\name{statistical_databases}
\alias{statistical_databases}
\title{Get List of All Databases for a given Theme from TUIK}
\usage{
statistical_databases(theme)
}
\arguments{
\item{theme}{Data Theme}
}
\value{
A data tibble
}
\description{
Get List of All Databases for a given Theme from TUIK
}
\examples{
\dontrun{statistical_databases(102)}

}
