% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuik-statistical-tables.R
\name{statistical_tables}
\alias{statistical_tables}
\title{Get List of All Statistical Tables for a given Theme from TUIK}
\usage{
statistical_tables(theme)
}
\arguments{
\item{theme}{Data Theme}
}
\value{
A data tibble
}
\description{
Get List of All Statistical Tables for a given Theme from TUIK
}
\examples{
\dontrun{statistical_tables(102)}

}
