% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuik-statistical-themes.R
\name{statistical_themes}
\alias{statistical_themes}
\title{Get Statistical Themes and URLS from TUIK}
\usage{
statistical_themes()
}
\value{
A tibble
}
\description{
Get Statistical Themes and URLS from TUIK
}
\examples{
\dontrun{
statistical_themes()
}

}
