% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuik-package.R
\docType{package}
\name{tuik-package}
\alias{tuik}
\alias{tuik-package}
\title{tuik: Download Data File And Database Urls from TUIK}
\description{
Download Data File And Database Urls from TUIK.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emraher/tuik}
  \item Report bugs at \url{https://github.com/emraher/tuik/issues}
}

}
\author{
\strong{Maintainer}: Emrah Er \email{eer@eremrah.com}

}
\keyword{internal}
